requirements.txt
galaxy.plugin.api==0.66.0
send2trash==1.5.0
requests==2.24.0
galaxyutils==0.1.5
